requirements.txt
streamlit
openai
PyPDF2
faiss-cpu
scikit-learn